+++ packages/react/src/index.ts
@@ -1,9 +1,9 @@
 import { router as Router } from '@inertiajs/core'
 
 export const router = Router
-export { default as createInertiaApp } from './createInertiaApp'
 export { default as Head } from './Head'
-export { default as Link, InertiaLinkProps } from './Link'
+export { InertiaLinkProps, default as Link } from './Link'
+export { default as createInertiaApp } from './createInertiaApp'
 export { default as useForm } from './useForm'
 export { default as usePage } from './usePage'
 export { default as useRemember } from './useRemember'