+++ packages/svelte/src/lib/store.ts
@@ -1,0 +1,17 @@
+import type { Page } from '@inertiajs/core'
+import { writable } from 'svelte/store'
+import type { ResolvedComponent } from './types'
+
+interface Store {
+  component: ResolvedComponent | null
+  page: Page | null
+  key?: number | null
+}
+
+const store = writable<Store>({
+  component: null,
+  page: null,
+  key: null,
+})
+
+export default store