+++ packages/vue2/src/useForm.ts
@@ -1,9 +1,11 @@
-import { Method, Progress, router, VisitOptions } from '@inertiajs/core'
+import { FormDataConvertible, Method, Progress, router, VisitOptions } from '@inertiajs/core'
 import cloneDeep from 'lodash.clonedeep'
 import isEqual from 'lodash.isequal'
-import Vue from 'vue'
+import { reactive, watch } from 'vue'
 
-interface InertiaFormProps<TForm> {
+type FormDataType = object
+
+interface InertiaFormProps<TForm extends FormDataType> {
   isDirty: boolean
   errors: Record<keyof TForm, string>
   hasErrors: boolean
@@ -14,8 +16,8 @@ interface InertiaFormProps<TForm> {
   data(): TForm
   transform(callback: (data: TForm) => object): this
   defaults(): this
-  defaults(field: keyof TForm, value: string): this
-  defaults(fields: Record<keyof TForm, string>): this
+  defaults(field: keyof TForm, value: FormDataConvertible): this
+  defaults(fields: Partial<TForm>): this
   reset(...fields: (keyof TForm)[]): this
   clearErrors(...fields: (keyof TForm)[]): this
   setError(field: keyof TForm, value: string): this
@@ -29,26 +31,29 @@ interface InertiaFormProps<TForm> {
   cancel(): void
 }
 
-export type InertiaForm<TForm> = TForm & InertiaFormProps<TForm>
+export type InertiaForm<TForm extends FormDataType> = TForm & InertiaFormProps<TForm>
 
 export interface InertiaFormTrait {
-  form<TForm>(data: TForm): InertiaForm<TForm>
-  form<TForm>(rememberKey: string, data: TForm): InertiaForm<TForm>
+  form<TForm extends FormDataType>(data: TForm): InertiaForm<TForm>
+  form<TForm extends FormDataType>(rememberKey: string, data: TForm): InertiaForm<TForm>
 }
 
-export default function useForm<TForm>(data: TForm): InertiaForm<TForm>
-export default function useForm<TForm>(rememberKey: string, data: TForm): InertiaForm<TForm>
-export default function useForm<TForm>(...args): InertiaForm<TForm> {
+export default function useForm<TForm extends FormDataType>(data: TForm | (() => TForm)): InertiaForm<TForm>
+export default function useForm<TForm extends FormDataType>(
+  rememberKey: string,
+  data: TForm | (() => TForm),
+): InertiaForm<TForm>
+export default function useForm<TForm extends FormDataType>(...args): InertiaForm<TForm> {
   const rememberKey = typeof args[0] === 'string' ? args[0] : null
   const data = (typeof args[0] === 'string' ? args[1] : args[0]) || {}
   const restored = rememberKey ? (router.restore(rememberKey) as { data: any; errors: any }) : null
-  let defaults = cloneDeep(data)
+  let defaults = typeof data === 'object' ? cloneDeep(data) : cloneDeep(data())
   let cancelToken = null
   let recentlySuccessfulTimeoutId = null
   let transform = (data) => data
 
-  const form = Vue.observable({
-    ...(restored ? restored.data : data),
+  const form = reactive({
+    ...(restored ? restored.data : cloneDeep(defaults)),
     isDirty: false,
     errors: restored ? restored.errors : {},
     hasErrors: false,
@@ -57,7 +62,7 @@ export default function useForm<TForm>(...args): InertiaForm<TForm> {
     wasSuccessful: false,
     recentlySuccessful: false,
     data() {
-      return Object.keys(data).reduce((carry, key) => {
+      return Object.keys(defaults).reduce((carry, key) => {
         carry[key] = this[key]
         return carry
       }, {})
@@ -68,6 +73,10 @@ export default function useForm<TForm>(...args): InertiaForm<TForm> {
       return this
     },
     defaults(key, value) {
+      if (typeof data === 'function') {
+        throw new Error('You cannot call `defaults()` when using a function to define your form data.')
+      }
+
       if (typeof key === 'undefined') {
         defaults = this.data()
       } else {
@@ -77,19 +86,18 @@ export default function useForm<TForm>(...args): InertiaForm<TForm> {
       return this
     },
     reset(...fields) {
-      let clonedDefaults = cloneDeep(defaults)
+      const resolvedData = typeof data === 'object' ? cloneDeep(defaults) : cloneDeep(data())
+      const clonedData = cloneDeep(resolvedData)
       if (fields.length === 0) {
-        Object.assign(this, clonedDefaults)
+        defaults = clonedData
+        Object.assign(this, resolvedData)
       } else {
-        Object.assign(
-          this,
-          Object.keys(clonedDefaults)
-            .filter((key) => fields.includes(key))
-            .reduce((carry, key) => {
-              carry[key] = clonedDefaults[key]
-              return carry
-            }, {}),
-        )
+        Object.keys(resolvedData)
+          .filter((key) => fields.includes(key))
+          .forEach((key) => {
+            defaults[key] = clonedData[key]
+            this[key] = resolvedData[key]
+          })
       }
 
       return this
@@ -225,20 +233,16 @@ export default function useForm<TForm>(...args): InertiaForm<TForm> {
     },
   })
 
-  new Vue({
-    created() {
-      this.$watch(
-        () => form,
-        (newValue) => {
-          form.isDirty = !isEqual(form.data(), defaults)
-          if (rememberKey) {
-            router.remember(newValue.__remember(), rememberKey)
-          }
-        },
-        { immediate: true, deep: true },
-      )
+  watch(
+    form,
+    (newValue) => {
+      form.isDirty = !isEqual(form.data(), defaults)
+      if (rememberKey) {
+        router.remember(cloneDeep(newValue.__remember()), rememberKey)
+      }
     },
-  })
+    { immediate: true, deep: true },
+  )
 
   return form
 }